move Parameter and child classes to their own file to prepare for pipeline (#63)
